Many changes and "improvements". Read description.
+ Improved build file by making the formatting more consistent and removing redundant keywords.
+ Converted from ModCore method to custom methods.
+ Changed method and class names to use "better" naming conventions.
- Eradicated ModCore.
