+++ build.gradle
@@ -1,73 +1,74 @@
 plugins {
-    id "com.github.johnrengelman.shadow" version "6.1.0"
-    id "net.minecraftforge.gradle.forge" version "b13624b"
+    id 'net.minecraftforge.gradle.forge' version '6f5327738d'
+    id 'com.github.johnrengelman.shadow' version '6.1.0'
+    id 'java'
 }
 
-version = project.mod_ver
-group = project.mod_group
-archivesBaseName = project.mod_name
+version = mod_ver
+group = mod_group
+archivesBaseName = mod_name
 
 sourceCompatibility = targetCompatibility = 1.8
 compileJava.options.encoding = 'UTF-8'
 
 minecraft {
     version = '1.8.9-11.15.1.2318-1.8.9'
-    runDir = "run"
-    replace('@VER@', project.version)
+    runDir = 'run'
+    replace('@VER@', mod_ver)
     mappings = 'stable_22'
-    makeObfSourceJar = false
+}
+
+configurations {
+    shade
+    implementation.extendsFrom(shade)
 }
 
 repositories {
     maven {
-        name = "JitPack"
-        url = "https://jitpack.io"
+        name = 'JitPack'
+        url = 'https://jitpack.io'
+    }
+    maven {
+        name = 'Sk1er'
+        url = 'https://repo.sk1er.club/repository/maven-public'
     }
 }
 
 dependencies {
-    provided files("libs/modcore.jar")
-    shadow("com.github.jncrmx:discord-game-sdk4j:0.5.3")
+    shade('com.github.jncrmx:discord-game-sdk4j:0.5.3')
+    shade('gg.essential:Vigilance:123-10809-SNAPSHOT')
 }
 
 jar {
+    duplicatesStrategy = DuplicatesStrategy.EXCLUDE
     manifest.attributes(
-            "FMLCorePlugin": "io.github.dediamondpro.hycord.tweaker.FMLLoadingPlugin",
-            "FMLCorePluginContainsFMLMod": true,
-            "ForceLoadAsMod": true,
-            "ModSide": "CLIENT",
-            "TweakOrder": "0",
+            "ModSide": "CLIENT"
     )
+    exclude 'META-INF/*.RSA', 'META-INF/*.SF', 'META-INF/*.DSA'
 }
 
 shadowJar {
-    configurations = [project.configurations.shadow]
-    duplicatesStrategy DuplicatesStrategy.EXCLUDE
+    configurations = [project.configurations.shade]
+    duplicatesStrategy = DuplicatesStrategy.EXCLUDE
 
-    classifier '' // Replace the default JAR
+    archiveClassifier.set('')
 }
 
 processResources {
-    // this will ensure that this task is redone when the versions change.
-    inputs.property "version", project.version
-    inputs.property "mcversion", project.minecraft.version
-
-    // replace stuff in mcmod.info, nothing else
+    inputs.property 'version', project.version
+    inputs.property 'mcversion', project.minecraft.version
+    inputs.property 'name', mod_name
     from(sourceSets.main.resources.srcDirs) {
-        include "mcmod.info"
-
-        // replace version and mcversion
-        expand "version": project.version, "mcversion": project.minecraft.version
+        include 'mcmod.info'
+        expand (
+                'version':project.version,
+                'mcversion':project.minecraft.version,
+                'name':mod_name
+        )
     }
-
-    // copy everything else, thats not the mcmod.info
     from(sourceSets.main.resources.srcDirs) {
-        exclude "mcmod.info"
+        exclude 'mcmod.info'
     }
-}
-
-reobf {
-    shadowJar {}
 }
 
 task moveResources {
@@ -79,5 +80,4 @@ task moveResources {
 
 tasks.moveResources.dependsOn processResources
 tasks.classes.dependsOn moveResources
-
 tasks.reobfJar.dependsOn(shadowJar)
