+++ gradle.properties
@@ -1,5 +1,5 @@
 org.gradle.jvmargs=-Xmx4G
 
 mod_name=HyCord
-mod_ver=1.2.0-pre5
+mod_ver=1.2.0-pre6
 mod_group=io.github.dediamondpro.hycord
