+++ src/main/java/io/github/dediamondpro/hycord/core/NetworkUtils.java
@@ -21,7 +21,7 @@ public class NetworkUtils {
             con.setRequestMethod("GET");
             int status = con.getResponseCode();
             if (status != 200) {
-                System.out.println("Api request failed, status code " + status);
+                System.out.println("API request failed, status code " + status);
                 return null;
             }
             BufferedReader in = new BufferedReader(new InputStreamReader(con.getInputStream()));
@@ -39,18 +39,18 @@ public class NetworkUtils {
         return null;
     }
 
-    public static String getUUID(String username) {
+    public static String getUuid(String username) {
         JsonElement response = getRequest("https://api.mojang.com/users/profiles/minecraft/" + username);
         if (response == null)
             return null;
         return response.getAsJsonObject().get("id").getAsString();
     }
 
-    public static BufferedImage getImg(String imgUrl) {
+    public static BufferedImage getImage(String imgUrl) {
         try {
             URL url = new URL(imgUrl);
             HttpURLConnection con = (HttpURLConnection) url.openConnection();
-            con.addRequestProperty("User-Agent", "HyCord");
+            con.addRequestProperty("User-Agent", "HyCord (Mozilla Firefox)");
             return ImageIO.read(con.getInputStream());
         } catch (IOException e) {
             e.printStackTrace();
