+++ src/main/java/io/github/dediamondpro/hycord/core/SimpleCommand.java
@@ -5,7 +5,7 @@ import net.minecraft.command.CommandException;
 import net.minecraft.command.ICommandSender;
 
 /**
- * Taken from NotEnoughUpdates under Attribution-NonCommercial 3.0 license
+ * Adapted from NotEnoughUpdates under Attribution-NonCommercial 3.0 license
  * https://github.com/Moulberry/NotEnoughUpdates/blob/master/LICENSE
  *
  * @author MoulBerry
