+++ src/main/java/io/github/dediamondpro/hycord/features/UpdateChecker.java
@@ -1,24 +1,20 @@
 package io.github.dediamondpro.hycord.features;
 
-import club.sk1er.mods.core.gui.notification.Notifications;
 import com.google.gson.JsonElement;
 import io.github.dediamondpro.hycord.HyCord;
 import io.github.dediamondpro.hycord.core.NetworkUtils;
 import io.github.dediamondpro.hycord.options.Settings;
-import kotlin.Unit;
-import net.minecraft.client.gui.GuiMainMenu;
-import net.minecraftforge.client.event.GuiOpenEvent;
-import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.common.event.FMLLoadCompleteEvent;
+import net.minecraft.client.Minecraft;
+import net.minecraft.event.ClickEvent;
+import net.minecraft.util.ChatComponentText;
+import net.minecraft.util.ChatStyle;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
-
-import java.awt.*;
-import java.io.IOException;
-import java.net.URI;
+import net.minecraftforge.fml.common.gameevent.TickEvent;
 
 public class UpdateChecker {
 
     public static JsonElement latest;
+    private static boolean sent;
 
     public static boolean checkUpdate() {
         JsonElement data = NetworkUtils.getRequest("https://api.github.com/repos/dediamondpro/hycord/releases");
@@ -37,21 +33,10 @@ public class UpdateChecker {
     }
 
     @SubscribeEvent
-    public void onGuiOpen(GuiOpenEvent e) {
-        if (!(e.gui instanceof GuiMainMenu)) return;
-        try {
-            Notifications.INSTANCE.pushNotification("Hycord version " + latest.getAsJsonObject().get("tag_name").getAsString() + " is available", "Click here to open GitHub", UpdateChecker::openTab);
-        } catch (Exception ex) {
-            ex.printStackTrace();
-        }
+    public void onClientTick(TickEvent.ClientTickEvent event) {
+        if (event.phase != TickEvent.Phase.START || sent) return;
+        Minecraft.getMinecraft().ingameGUI.getChatGUI().printChatMessage(new ChatComponentText("HyCord version " + latest.getAsJsonObject().get("tag_name").getAsString() + " is available. Click here to open GitHub").setChatStyle(new ChatStyle().setChatClickEvent(new ClickEvent(ClickEvent.Action.OPEN_URL, latest.getAsJsonObject().get("html_url").getAsString()))));
+        sent = true;
     }
 
-    private static Unit openTab() {
-        try {
-            Desktop.getDesktop().browse(URI.create(latest.getAsJsonObject().get("html_url").getAsString()));
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return Unit.INSTANCE;
-    }
 }
