+++ src/main/java/io/github/dediamondpro/hycord/options/Settings.java
@@ -1,12 +1,13 @@
 package io.github.dediamondpro.hycord.options;
 
-import club.sk1er.vigilance.Vigilant;
-import club.sk1er.vigilance.data.Property;
-import club.sk1er.vigilance.data.PropertyType;
+import gg.essential.vigilance.Vigilant;
+import gg.essential.vigilance.data.Property;
+import gg.essential.vigilance.data.PropertyType;
 
 import java.io.File;
 
 public class Settings extends Vigilant {
+
     @Property(
             type = PropertyType.SELECTOR,
             name = "Update Channel",
